+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+* Allow export of prfKey, IdCredSec, blinding randomness and attribute commitment randomness using the new paths.
+
 ## 4.0.1
 
 ### Changed